fix (flexible): correct summarizing flexible component's priorities.
Use `vim.tbl_contains` to ensure a value is not already in the
`priorties` list. `priorities[priority]` may return erroneous answers.
For example, if the first component has a priority value of `2` and the
second component has a priority value of `1`, `priorities[1]` returns
the priority of the first component (2) and not `nil` as expected. This
results in ignoring the second component.